Split context to the separate file
